converted project from float to double

+++ NeuroSharp/Layers/ActivationLayer.cs
@@ -5,28 +5,28 @@ namespace NeuroSharp
 {
     public class ActivationLayer : Layer
     {
-        public Func<float, float> Activation { get; set; }
-        public Func<float, float> ActivationPrime { get; set; }
+        public Func<double, double> Activation { get; set; }
+        public Func<double, double> ActivationPrime { get; set; }
 
-        public ActivationLayer(Func<float, float> activation, Func<float, float> activationPrime)
+        public ActivationLayer(Func<double, double> activation, Func<double, double> activationPrime)
         {
             Activation = activation;
             ActivationPrime = activationPrime;
         }
 
-        public override Vector<float> ForwardPropagation(Vector<float> input)
+        public override Vector<double> ForwardPropagation(Vector<double> input)
         {
             Input = input;
-            Output = Vector<float>.Build.Dense(input.Count);
+            Output = Vector<double>.Build.Dense(input.Count);
             for(int i = 0; i < input.Count; i++)
                 Output[i] = Activation(input[i]);
 
             return Output;
         }
 
-        public override Vector<float> BackPropagation(Vector<float> outputError, OptimizerType optimzerType, int sampleIndex, float learningRate = 0.001f)
+        public override Vector<double> BackPropagation(Vector<double> outputError, OptimizerType optimzerType, int sampleIndex, double learningRate = 0.001f)
         {
-            Vector<float> vec = Vector<float>.Build.Dense(Input.Count);
+            Vector<double> vec = Vector<double>.Build.Dense(Input.Count);
             for (int i = 0; i < Input.Count; i++)
                 vec[i] = ActivationPrime(Input[i]);
             return vec.PointwiseMultiply(outputError);
