+++ NeuroSharp/Layers/SoftmaxActivationLayer.cs
@@ -5,16 +5,16 @@ namespace NeuroSharp
 {
     public class SoftmaxActivationLayer : Layer
     {
-        public override Vector<float> ForwardPropagation(Vector<float> input)
+        public override Vector<double> ForwardPropagation(Vector<double> input)
         {
             Input = input;
             Output = ActivationFunctions.Softmax(input);
             return Output;
         }
 
-        public override Vector<float> BackPropagation(Vector<float> outputError, OptimizerType optimzerType, int sampleIndex, float learningRate = 0.001f)
+        public override Vector<double> BackPropagation(Vector<double> outputError, OptimizerType optimzerType, int sampleIndex, double learningRate = 0.001f)
         {
-            Matrix<float> jacobian = ActivationFunctions.SoftmaxPrime(Input);
+            Matrix<double> jacobian = ActivationFunctions.SoftmaxPrime(Input);
             return jacobian * outputError;
         }
     }
